+++ tests/testthat/test-sclr.R
@@ -14,6 +14,10 @@ test_that("Error with missing parameters", {
     expect_error(sclr(one_titre_data))
 })
 
+test_that("Warning with no covariates", {
+  expect_warning(sclr(status ~ 1, one_titre_data))
+})
+
 test_that("Error with unexpected outcome", {
     # Factor
     dat <- one_titre_data
@@ -26,7 +30,7 @@ test_that("Error with unexpected outcome", {
 
 test_that("Returns the expected parameter names", {
   fit_my_names <- sclr(status ~ logHI, one_titre_data)
-  expect_named(fit_my_names$parameters, c("lambda", "beta_0", "beta_logHI"))
+  expect_named(fit_my_names$parameters, c("theta", "beta_0", "beta_logHI"))
   est_conv_names <- sclr(
     status ~ logHI, one_titre_data, conventional_names = TRUE
   )